DB CRUD operations with php & page response
I forgot to commit all changes and updates

Most of the code is from the previous project of
CRUD operations with PHP and Javascript.

The contents of the upload folder will be generated
during runtime and may differ for each instance of
the project. I think I should make it so that the
upload folder will be created if there is no existing
folder.

The read,add,and update operations are the only ones
available for now.

I did not create an implementation
for the delete operation since the operation was not
specified with the project requirement.