Image upload preparation & xdebug config
I don't get how xdebug works in my local computer
but won't work when installed in the lab computer.

I don't understand where the setting for the xdebug
is set-up on.

The contents of the .vscode folder may be of use when
cloning the repo to reproduce the project environment
on other machines.