User page html & js form
The user page displays all the registered users.

Every new registration and user update during session,
the images for the user profile picture is from the
local client.

The image for pfp during user registration and update
is reused for local display.

I think the function for handling the ajax request to the
database is still fucked. There are portions that need to
be fixed.

But it works somehow.

The input validation is poorly implemented, just properly
identifying the empty inputs would be fine for now to
keep up with the project requirements.

I think this won't work with future updates.