Register & Update invoke the same request function
The functions to send data to the server for
the user update and registration
have the same functionality and structure.

Both differ in values for certain function calls,
by making it so that the variables for those function
calls come from parameters.

It is possible to use the same function to send
data with differing parameters.